Fix default security object handling
There's a race where if a database is opened with a default_security set
and it crashes before first compact, and is then reopened after the
default_security option has changed that it will pick the second
security option. This ensures change closes that relatively obscure bug
that was only found during testing.