[SQUERGE] couch_db_purge_docs_tests cleanup
I've taken some liberty cleaning up the tests for purges in this module
by using a few helper functions and some minor reformatting. Currently
the repeated docid test is failing because there's a bug in the
purge_docs logic in couch_db_updater where its not accounting for the
fact that a user may have specified a doc id multiple times.

As for doc updates we'll have to apply purges from the first purge infos
and return responses that correspond to which updates actually took
effect.