Check security context inside transactions
Previously security context was checked in a separate transaction, even when
the db handle was fetched from the cache. That preserved the same behavior as
CouchDB < 4.0 where database "open" operation would fail with an authorization
error. However, it turns out to be pretty inneficient.

To fix the issue, mark the db handle as needing a membership check after it is
fetched from the cache and opened but don't actually perform the check. Then,
when the main transaction is run, the actual check if preformed if it was
marked as such.